[Finishes #157721787] Report scores in GameResult with mean 6.5 for more typical scoring

+++ play_hearts/server/PlayerSession.h
@@ -47,6 +47,6 @@ private:
   std::string mSessionToken;
 
   GameState* mGameState;
-  float mTotals[4] = {0};
-  float mReferenceTotals[4] = {0};
+  int mTotals[4] = {0};
+  int mReferenceTotals[4] = {0};
 };
